Fixed message board no accessible if logged-out
